Update why-tcphol.md
Minor language suggestions - "less and less good" is incorrect grammar I think; we could say "less and less well" or "more and more poorly", but both I think are more awkward than simply saying the "performance gets worse". or for emphasis "the performance gets worse and worse".

I also removed the word "up" later in the sentence which didn't seem to be quite correct. I wanted to change this sentence to:

> "...they typically have six TCP connections over which the last packet may be distributed so for each lost packet the other connections without loss can still continue."

but I wasn't sure if had understood the meaning of that sentence correctly.